requirements.txt
streamlit==1.32.0
openai==1.12.0
pymongo==4.6.2
python-dotenv==1.0.1
httpx==0.24.1
langchain==0.1.12
python-magic==0.4.27
graphviz==0.20.1
cryptography==42.0.5 